+++ src/index.css
@@ -45,60 +45,60 @@
       @apply text-base not-italic text-gray2;
     }
     .ql-container,
-    .ql-editor p[data-value="medium"]::before {
+    .ql-editor p[data-value='medium']::before {
       @apply text-base;
     }
 
-    #customDatePicker {
-      .react-datepicker-wrapper,
-      .react-datepicker__triangle {
-        @apply hidden
-      }
-      .react-datepicker-popper {
-        @apply transform-none top-auto mt-1 !important
-      }
-      .react-datepicker {
-        @apply shadow-modal border-none rounded
-      }
-      .react-datepicker__day-name{
-        @apply text-white
-      }
-      .react-datepicker__header {
-        @apply bg-oliveGreen1 text-white border-none relative text-center
-      }
-      .react-datepicker__day--selected {
-        @apply bg-oliveGreen1 text-white rounded-lg
-      }
-     }
-
-    .announcement-text-divider::after {
-      content: '|';
-      margin: 14px;
-      color: #b5b5b5;
-    }
-    .lounge-text-divider::after {
-      content: '|';
-      margin: 8px;
-      color: #b5b5b5;
-    }
-    .meal-recruit-text-divider::after {
-      content: '|';
-      margin: 8px;
-      color: #6d6d6d;
-    }
+    /* DatePicker CSS */
     .react-datepicker-wrapper {
-      width: 100%;
-      z-index: 0;
+      @apply relative z-0 h-full w-full;
+    }
+    .react-datepicker__input-container {
+      @apply h-full;
+    }
+    .react-datepicker__triangle {
+      @apply hidden;
     }
     .react-datepicker-popper {
-      z-index: 1000;
+      @apply z-20 m-0;
     }
-    .react-datepicker__day--selected {
-      background-color: #6fa235;
+    .react-datepicker {
+      @apply rounded-xl border-none shadow-modal;
+    }
+    .react-datepicker__day-name {
+      @apply text-white;
     }
     .react-datepicker__header {
-      background-color: #eaf8e6;
+      @apply rounded-t-xl border-none bg-oliveGreen1 py-2.5 text-white;
     }
+    .react-datepicker__header:not(.react-datepicker__header--has-time-select) {
+      @apply rounded-tr-xl;
+    }
+    .react-datepicker__day--selected {
+      @apply rounded-lg !bg-oliveGreen1 text-white;
+    }
+    .react-datepicker__day--keyboard-selected,
+    .react-datepicker__month-text--keyboard-selected,
+    .react-datepicker__quarter-text--keyboard-selected,
+    .react-datepicker__year-text--keyboard-selected {
+      @apply bg-oliveGreen1 text-white;
+    }
+  }
+
+  .announcement-text-divider::after {
+    content: '|';
+    margin: 14px;
+    color: #b5b5b5;
+  }
+  .lounge-text-divider::after {
+    content: '|';
+    margin: 8px;
+    color: #b5b5b5;
+  }
+  .meal-recruit-text-divider::after {
+    content: '|';
+    margin: 8px;
+    color: #6d6d6d;
   }
 }
